requirements.txt
streamlit==1.32.0
spacy==3.7.4
pandas==2.2.1
networkx==3.2.1
plotly==5.19.0
transformers==4.38.2
torch==2.2.1
sentence-transformers==2.6.1
numpy==1.26.4
scikit-learn==1.4.1.post1

# Document processing libraries
PyPDF2==3.0.1
pdfplumber==0.10.0
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2

# OCR libraries
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0

# Advanced Analytics libraries
folium==0.15.1
streamlit-folium==0.15.0
textblob==0.17.1
vaderSentiment==3.3.2
geopy==2.4.1
pycountry==23.12.11
dateparser==1.2.0
wordcloud==1.9.2